+++ fixtures/react_native_windows_e2e_app/scripts/VSComponentList.ps1
@@ -22,4 +22,4 @@ do
 
 Get-Content $dir\err
 Get-Content $dir\out
-Get-Content $vsconfig 
+Get-Content $vsconfig
