+++ packages/haul-core/src/webpack/plugins/RamBundlePlugin/__tests__/__fixtures__/nestedCrashFn.esm.js
