+++ packages/haul-core/src/webpack/plugins/RamBundlePlugin/index.ts
@@ -1,0 +1,1 @@
+export { RamBundlePlugin } from './RamBundlePlugin';
