+++ packages/haul-core/src/webpack/plugins/RamBundlePlugin/minifyWorker.ts
