--- packages/haul-core/src/webpack/plugins/RamBundleWebpackPlugin/index.ts
@@ -1,1 +1,0 @@
-export { default } from './RamBundleWebpackPlugin';
