--- packages/haul-core/src/webpack/plugins/RamBundleWebpackPlugin/utils.ts
@@ -1,2 +1,0 @@
-export const countLines = (string: string) =>
-  (string.match(/\r\n?|\n|\u2028|\u2029/g) || []).length;
