+++ packages/haul-core/src/webpack/plugins/__tests__/__fixtures__/async.js
@@ -1,1 +1,3 @@
+'use strict';
+
 export default () => console.log('async');
