+++ packages/haul-core/src/webpack/plugins/__tests__/__fixtures__/entry.js
@@ -1,3 +1,5 @@
+'use strict';
+
 console.log('entry');
 
 import('./async').then(m => m.default());
