+++ packages/haul-core/src/webpack/plugins/__tests__/__fixtures__/index.js
@@ -1,5 +1,7 @@
 'use strict';
 
+import './async';
+
 const str = 'string';
 
 str.someMethod = function() {}
